+++ test/transform/resource/before/BuilderToBuilderMethodName.java
@@ -1,0 +1,7 @@
+import lombok.Builder;
+
+@Builder(toBuilder = true, toBuilderMethodName = "mutate")
+class BuilderToBuilderMethodName {
+	private String name;
+	private int age;
+}